+++ src/Components/Atoms/Atoms.tsx
@@ -1,108 +1,4 @@
-// import { css } from "@emotion/react";
 import styled from "@emotion/styled";
-import { GoSearch } from "react-icons/go";
-import { BsFacebook } from "react-icons/bs";
-import { AiFillHome, AiOutlineRight } from "react-icons/ai";
-import { MdOndemandVideo } from "react-icons/md";
-import { AiOutlineShop } from "react-icons/ai";
-import { HiUserGroup } from "react-icons/hi";
-import { BiCreditCard } from "react-icons/bi";
-import { CgMenuGridO } from "react-icons/cg";
-import { RiMessengerFill } from "react-icons/ri";
-import { IoIosNotifications } from "react-icons/io";
-import { RiSettings5Fill, RiFeedbackFill } from "react-icons/ri";
-import { MdOutlineHelp } from "react-icons/md";
-import { FaMoon } from "react-icons/fa";
-import { GrLogout } from "react-icons/gr";
-
-export const SearchIcon = styled(GoSearch)`
-  color: #67656b;
-  font-size: 15px;
-  left: 5%;
-  position: absolute;
-`;
-
-export const Facebook = styled(BsFacebook)`
-  color: #1877f2;
-  font-size: 45px;
-`;
-
-export const Home = styled(AiFillHome)`
-  color: #1877f2;
-  cursor: pointer;
-  font-size: 23px;
-  font-weight: 400;
-`;
-
-export const Video = styled(MdOndemandVideo)`
-  color: #67656b;
-  cursor: pointer;
-  font-size: 23px;
-`;
-
-export const Shop = styled(AiOutlineShop)`
-  color: #67656b;
-  cursor: pointer;
-  font-size: 23px;
-`;
-
-export const Groupes = styled(HiUserGroup)`
-  color: #67656b;
-  cursor: pointer;
-  font-size: 23px;
-`;
-
-export const Feeds = styled(BiCreditCard)`
-  color: #67656b;
-  cursor: pointer;
-  font-size: 23px;
-`;
-
-export const Menu = styled(CgMenuGridO)`
-  cursor: pointer;
-  font-size: 23px;
-`;
-
-export const Messenger = styled(RiMessengerFill)`
-  cursor: pointer;
-  font-size: 23px;
-`;
-
-export const Notifications = styled(IoIosNotifications)`
-  cursor: pointer;
-  font-size: 23px;
-`;
-
-export const Settings = styled(RiSettings5Fill)`
-  cursor: pointer;
-  font-size: 23px;
-`;
-
-export const Support = styled(MdOutlineHelp)`
-  cursor: pointer;
-  font-size: 23px;
-`;
-
-export const Display = styled(FaMoon)`
-  cursor: pointer;
-  font-size: 23px;
-`;
-
-export const FeedBack = styled(RiFeedbackFill)`
-  cursor: pointer;
-  font-size: 23px;
-`;
-
-export const LogOut = styled(GrLogout)`
-  color: #67656b;
-  cursor: pointer;
-  font-size: 20px;
-`;
-
-export const RightArrow = styled(AiOutlineRight)`
-  cursor: pointer;
-  font-size: 23px;
-`;
 
 export const Ptag = styled("p")`
   font-size: 14px;
